Updates 'inline and defer CSS' to be compatible with Cloudflare
Hello, while trying 'inline and defer CSS', I found that the inline CSS is working, but the full CSS is not being deffered correctly (it's not even loading..). So, while trying to understand the bug, I remember myself that I have CloudFlare's Rocket Loader enabled and..for wtf reasons, I found that Cloudflare is trying to 'Rocketize' the inline script that loads the CSS after the page is loaded. But..there's no script to load asynchronously, or the script is not compatible with Rocket Loader...well, I don't know, but anyway, because the inline script is so small, I just added a data tag to remember CloudFlare to not modify the inline script, and so the CSS loads and we will be all happy. \o/