Optimize Images (#160)
"optimize images" is a new option in "Extra" which will route requests for images (img src, srcsets, css background images and data-thumb attribs) through an image optimizing proxy (in this case provided by Shortpixel but this can be changed through a filter).

The service is provided for free by Shortpixel during the AO beta-cycle and will remain free up until a still to be defined threshold per domain after the beta. If the threshold is reached extra credits will have to be purchased at shortpixel.com.