Added '.min' filter option
The filter option is by default set to false, so it will not break the current check validation.

REMARK: For a futher release it would be nice that we can set the default to true. This would prevent then by default that other minification software will NOT touch the already minified files from autoptimize.